feat: remove obsolete `MathParser_min.lua` file
This commit removes the obsolete `MathParser_min.lua` file from the
repository. The minified version of the MathParser script is no longer
needed as we have transitioned to using the unminified version for
better readability and maintainability.

The `MathParser_min.lua` file was originally included to reduce the
file size for deployment purposes. However, I've decided that it would
be a better idea to move it to the GitHub releases page instead.

Signed-off-by: bytexenon <danielgtdavis@proton.me>