refactor: update author information and dates in source files
This commit updates the author information and dates in several source
files. The previous author information included an alias and a name,
which has now been simplified to just the alias "ByteXenon". Additionally,
the dates have been updated to reflect the current modification date.

The changes are made to improve consistency and clarity in the codebase.
The updated author information and dates are as follows:
- Author: ByteXenon
- Date: 2024-10-20

The affected files include:
- MathParser.lua
- src/Evaluator/Evaluator.lua
- src/Helpers/Helpers.lua
- src/Lexer/Lexer.lua
- src/Lexer/TokenFactory.lua
- src/MathParser.lua
- src/Parser/NodeFactory.lua
- src/Parser/Parser.lua

Signed-off-by: bytexenon <danielgtdavis@proton.me>