feat: add function to insert values from one table to another in Helpers.lua
This commit adds a new function `Helpers.insertValues` to the Helpers
module. The function allows for inserting values from a source table
into a destination table, with a check to ensure that keys do not
already exist in the destination table.

The new function provides a useful utility for mimicking OOP in the
future where all classes will have a `<ClassName>Methods` table that will be
merged into the class table.

Signed-off-by: bytexenon <danielgtdavis@proton.me>